Separate db query method from api request handling
